requirements.txt
# Core dependencies
anthropic>=0.28.0
openai>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# YAML configuration
PyYAML>=6.0

# Git integration
GitPython>=3.1.0

# Process execution and monitoring
psutil>=5.9.0

# HTTP requests
requests>=2.31.0
httpx>=0.24.0

# Logging and monitoring
structlog>=23.0.0
colorlog>=6.7.0

# Testing frameworks
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
isort>=5.12.0

# Notification channels
slack-sdk>=3.21.0

# Async support
aiofiles>=23.0.0
asyncio>=3.4.3

# Database (optional, for metrics storage)
sqlalchemy>=2.0.0
alembic>=1.11.0

# CLI
click>=8.1.0
rich>=13.0.0

# Utilities
python-dateutil>=2.8.0
tenacity>=8.2.0
